scripts/requirements.txt
sqlalchemy
fastapi
python-dotenv
psycopg